put ArbQueryParamProvider in useArbQueryParams file
