+++ packages/arb-token-bridge-ui/src/hooks/useArbQueryParams.tsx
@@ -13,9 +13,12 @@
     `setQueryParams(newAmount)`
 
 */
-
+import React from 'react'
+import { ReactRouter5Adapter } from 'use-query-params/adapters/react-router-5'
+import { parse, stringify } from 'query-string'
 import {
   NumberParam,
+  QueryParamProvider,
   StringParam,
   useQueryParams,
   withDefault
@@ -59,4 +62,24 @@ const AmountQueryParam = {
 
     return amountStr ?? ''
   }
+}
+
+export function ArbQueryParamProvider({
+  children
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <QueryParamProvider
+      adapter={ReactRouter5Adapter}
+      options={{
+        searchStringToObject: parse,
+        objectToSearchString: stringify,
+        updateType: 'replaceIn', // replace just a single parameter when updating query-state, leaving the rest as is
+        removeDefaultsFromUrl: true
+      }}
+    >
+      {children}
+    </QueryParamProvider>
+  )
 }
